// src/game/chess/type.ts
export enum Fen {
    WHITE_PAWN = "P",
    WHITE_KNIGHT = "N",
    WHITE_BISHOP = "B",
    WHITE_ROOK = "R",
    WHITE_QUEEN = "Q",
    WHITE_KING = "K",
    BLACK_PAWN = "p",
    BLACK_KNIGHT = "n",
    BLACK_BISHOP = "b",
    BLACK_ROOK = "r",
    BLACK_QUEEN = "q",
    BLACK_KING = "k",
}

export enum Material {
    DEFAULT = "d",
    SAPPHIRE = "s",
    AMBER = "a",
    GLASS = "l",
    STEAL = "t",
    GOLD = "g",
    PAPER = "p",
    EMERALD = "e",
}

export enum Base {
    DEFAULT = "d",
    SAPPHIRE = "s",
    AMBER = "a",
    GLASS = "l",
}